+++ src/Controller/SearchCardTabController.java
@@ -13,9 +13,13 @@ public class SearchCardTabController {
 
     @FXML
     private VBox BaseVBox;
+    
+    private int height = 272;
 
 
     public void SetVBox(List<Barang> temp) throws IOException {
+        int itemcount = temp.size();
+        int counter = 0;
         for(Barang barang : temp){
             FXMLLoader loader = new FXMLLoader();
             loader.setLocation(getClass().getResource("/GraphicUserInterface/SearchCard.fxml"));
@@ -24,6 +28,13 @@ public class SearchCardTabController {
             searchCardController.setCard(barang);
 
             BaseVBox.getChildren().add(cardbox);
+
+            
+            if (counter < itemcount-1){
+                BaseVBox.setPrefHeight(BaseVBox.getPrefHeight()+height);
+
+            }
+            counter++;
 
         }